+++ src/main/java/com/cosmian/jna/findex/structs/ToIndexedValue.java
@@ -1,0 +1,5 @@
+package com.cosmian.jna.findex.structs;
+
+public interface ToIndexedValue {
+    public IndexedValue toIndexedValue();
+}